Restructure module into a package.
